:rocket: Refactot project structure for better testing.
Add node-config package.
