+++ app/package.json
@@ -7,9 +7,11 @@
   },
   "dependencies": {
     "body-parser": "~1.17.1",
+    "config": "~1.26.1",
     "debug": "~2.6.3",
     "express": "~4.15.2",
     "express-validator": "~3.2.1",
+    "js-yaml": "~3.9.0",
     "moment": "~2.18.1",
     "mysql": "~2.13.0"
   }
